Added palette swap functionality using KoBeWi's "Godot Palette Swap Shader," found here: https://github.com/KoBeWi/Godot-Palette-Swap-Shader
Also added a "round_start" script, which currently sets up two different player characters with an os for each and a palette for each, as well as the camera. The parameters can be set in the editor for now, and later on we'll pass them via a character select screen.

Signed-off-by: Ethelyn <metallicminiboss@gmail.com>